+++ src/components/Pages/AppOptions/AppOptions.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+import "./AppOptions.scss";
+
+export const AppOptions: React.FC = () => {
+    return (
+        <>
+            AppOptions
+        </>
+    );
+};