+++ src/components/Pages/GamePage/GamePage.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+import "./GamePage.scss";
+
+export const GamePage: React.FC = () => {
+    return (
+        <>
+            GamePage
+        </>
+    );
+};