+++ src/components/Pages/RolesExplaination/RolesExplaination.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+import "./RolesExplaination.scss";
+
+export const RolesExplaination: React.FC = () => {
+    return (
+        <>
+            RolesExplaination
+        </>
+    );
+};