+++ src/components/Pages/RulesDescription/RulesDescription.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+import "./RulesDescription.scss";
+
+export const RulesDescription: React.FC = () => {
+    return (
+        <>
+            RulesDescription
+        </>
+    );
+};